requirements.txt
langchain==0.3.26
langchain-core==0.3.80
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
langchain-community==0.3.26
langchain-google-genai==2.0.8
google-ai-generativelanguage>=0.6.0,<0.7.0
google-generativeai>=0.8.0,<0.9.0
cachetools>=5.3.0,<6.0.0
tenacity>=8.1.0,<9.0.0
protobuf<5.0.0
rich<14
langchain-huggingface
pinecone
google-genai
youtube-transcript-api==0.6.2
-e .